archive test files and re-organize heuristics
