+++ python/heuristics/message_reasoning_labels.py
@@ -1,5 +1,5 @@
 import re
-from helpers_tuned import *
+from helpers import *
 
 class MessageReasoningLabels:
     def __init__(self, message, sender = None):
